Removed unnecessary checks when closing
The end result of the CloseCalendarFlyout() method always will be that the IsCalendarOpen in the for DateDiff_FromDate and DateDiff_ToDate will be false. For this, checking to see if they are true before making them false is redundant. They will be false when the method is called, no matter what.