Edit UnitConverterTestUnitTypeSwitching (Edit UnitConverterTest.cpp)
The test has been improved in the code, now it checks for changes after the function of SwitchActive and fits a new value separately into the to field after the function of SwitchActive
When I tried to edit the SwitchActive function, I removed the swap of fromType and toType. And I ran all the tests.
And only UnitConverterTestScientificInputs detected a value error when using the SwitchActive function.